% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{expand_function}
\alias{expand_function}
\title{Expands code within a function}
\usage{
expand_function(fun, envir = parent.frame())
}
\arguments{
\item{fun}{the function}

\item{envir}{the environment where to look for macros and other symbols}
}
\description{
Expands code within a function
}
