% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{onload}
\alias{onload}
\title{The onLoad hook}
\usage{
onload(pkg_name)
}
\arguments{
\item{pkg_name}{the package name}
}
\description{
It looks in the package code for functions and recursively
expands all macros. After expansion it compiles the function
to bytecode using the \code{compiler} package.
}
