% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{onload}
\alias{onload}
\title{The onLoad hook}
\usage{
onload(envir = topenv(parent.frame()))
}
\arguments{
\item{envir}{an environment in which the macro expansion is being done.}
}
\description{
It looks in the package code for functions and recursively
expands all macros. After expansion it compiles the function
to bytecode using the \code{compiler} package.
}
